multilingual-movies-recommendation-engine/requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
annoy==1.17.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autocommand==2.2.1
autoflake==1.7.6
backcall==0.2.0
beautifulsoup4==4.11.1
bertopic==0.12.0
bleach==5.0.1
blinker==1.5
bs4==0.0.1
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
cheroot==8.6.0
CherryPy==18.8.0
cinemagoer==2022.2.11
click==8.1.3
click-didyoumean==0.3.0
cohere==2.6.1
commonmark==0.9.1
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
Cython==0.29.32
datasets==2.6.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
docx==0.2.4
entrypoints==0.4
etils==0.8.0
evaluate==0.2.2
executing==1.1.1
fastjsonschema==2.16.2
fasttext==0.9.2
feedparser==6.0.10
filelock==3.8.0
fire==0.4.0
flatbuffers==22.10.26
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.11.0
gast==0.4.0
gcsfs==2022.11.0
gitdb==4.0.9
GitPython==3.1.29
google-api-core==2.10.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
grpcio==1.47.2
grpcio-tools==1.47.2
h5py==3.7.0
hdbscan==0.8.28
huggingface-hub==0.10.1
idna==3.4
IMDbPY==2022.7.9
importlib-metadata==5.0.0
importlib-resources==5.10.0
inflect==6.0.2
install==1.3.5
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
isort==5.10.1
jaraco.classes==3.2.3
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.10.0
jax==0.3.21
jaxlib==0.3.20
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mmh3==3.0.0
more-itertools==9.0.0
multidict==6.0.2
multiprocess==0.70.13
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.7
notebook==6.5.1
notebook_shim==0.2.0
numba==0.56.3
numpy==1.23.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pattern3==3.0.0
pdfminer.six==20220524
pdfminer3k==1.3.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
ply==3.11
podsearch==0.3.1
portend==3.1.0
progressbar==2.5
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.1
pycountry==22.3.5
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0b4
pyflakes==2.5.0
Pygments==2.13.0
pymilvus==2.1.3
Pympler==1.0.1
pynndescent==0.5.7
pyparsing==3.0.9
pyrate-limiter==2.8.3
pyrsistent==0.18.1
python-dateutil==2.8.2
pytrec-eval==0.5
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rich==12.6.0
rsa==4.9
scikit-learn==1.1.2
scikit-multilearn==0.2.0
scipy==1.9.2
seaborn==0.12.1
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
setfit==0.4.1
sgmllib3k==1.0.0
simplejson==3.17.6
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
stack-data==0.5.1
stqdm==0.0.4
streamlit==1.13.0
tempora==5.0.2
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.0
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.23.1
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
ujson==5.4.0
umap-learn==0.5.3
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
wrapt==1.14.1
xxhash==3.1.0
yapf==0.32.0
yarl==1.8.1
zc.lockfile==2.0
zipp==3.9.0
